+++ app/src/js/actions/index.js
@@ -176,12 +176,13 @@ export const getCumulusInstanceMetadata = () => ({
   }
 });
 
-export const getGranule = (granuleId) => ({
+export const getGranule = (granuleId, params) => ({
   [CALL_API]: {
     type: types.GRANULE,
     method: 'GET',
     id: granuleId,
-    path: `granules/${granuleId}`
+    path: `granules/${granuleId}`,
+    params
   }
 });
 
@@ -419,6 +420,11 @@ export const searchGranules = (infix) => ({ type: types.SEARCH_GRANULES, infix }
 export const clearGranulesSearch = () => ({ type: types.CLEAR_GRANULES_SEARCH });
 export const filterGranules = (param) => ({ type: types.FILTER_GRANULES, param });
 export const clearGranulesFilter = (paramKey) => ({ type: types.CLEAR_GRANULES_FILTER, paramKey });
+export const toggleGranulesTableColumns = (hiddenColumns, allColumns) => ({
+  type: types.TOGGLE_GRANULES_TABLE_COLUMNS,
+  hiddenColumns,
+  allColumns
+});
 
 export const getOptionsCollectionName = (options) => ({
   [CALL_API]: {